+++ ios/FluentUI/DrawerVnext/Drawer.swift
@@ -3,37 +3,46 @@
 //  Licensed under the MIT License.
 //
 
-import UIKit
 import SwiftUI
 
-@objc(DrawerState)
+// MARK: State
 
+/*
+ `DrawerState` assist to configure drawer functional properties via UIKit components.
+ */
+@objc(DrawerState)
 public class DrawerState: NSObject, ObservableObject {
-    weak var delegate: DrawerStateDelegate?
-    // presentation direction sets the direction for drawer slide out
-    @objc @Published public var presentationDirection: DrawerDirection = .left
-    // state to expand/collpase drawer
+
+    /// A callback executed when the drawer is expanded/collapsed
+    public var onStateChange: (DrawerStateChangedCompletionBlock)?
+
+    /*
+     Set `isExpanded` to `true` to maximize the drawer's width to fill the device screen horizontally minus the safe areas. Set to `false` to restore it to the normal size.
+     */
     @objc @Published public var isExpanded: Bool = false {
         didSet {
-            if isExpanded {
-                delegate?.drawerDidExpand?()
-            } else {
-                delegate?.drawerDidCollapsed?()
-            }
+            onStateChange?(isExpanded)
         }
     }
-    // if set to true background is dimmed or else by default is clear
+
+    @objc @Published public var presentationDirection: DrawerDirection = .left
+
+    /*
+     Set `backgroundDimmed` to `true` to dim the spacer area between drawer and base view. If set to `false` it restores to `clear` color
+     */
     @objc @Published public var backgroundDimmed: Bool = false
 }
 
-@objc(DrawerStateDelegate)
+public typealias DrawerStateChangedCompletionBlock = (_ isExpanded: Bool) -> Void
 
-// State delegate is currently required to notify hosting viewcontroller about expansion
-protocol DrawerStateDelegate: AnyObject {
-    @objc optional func drawerDidExpand()
-    @objc optional func drawerDidCollapsed()
+@objc(DrawerViewModel)
+public class DrawerAnimationModel: NSObject, ObservableObject {
+    @Published public var expand: Bool = false
 }
 
+/*
+ `DrawerTokens` assist to configure drawer apperance via UIKit components.
+ */
 public class DrawerTokens: ObservableObject {
 
     @Published public var shadowColor: Color!
@@ -55,24 +64,32 @@ public class DrawerTokens: ObservableObject {
     }
 }
 
-// Mark - Drawer View
-
 @objc public enum DrawerDirection: Int, CaseIterable {
+    /// Drawer animated right from a left base
     case left
+    /// Drawer animated right from a right base
     case right
 }
 
+// MARK: Drawer
+
 /**
  `Drawer` is used to present a content that is currently in hideout mode when collapsed and when expanded is used to present content view partially either on the left or right.
- 
- Currently `Drawer` only support horizontal axis and is expanded by default from left side of the screen unless explicitly specified
- 
+ `Drawer`  support horizontal axis and is expanded by default from left side of the screen unless explicitly specified
  Set `Content` to provide content for the drawer.
  */
 public struct Drawer<Content: View>: View {
 
-    @ObservedObject var state: DrawerState
-    @ObservedObject var tokens: DrawerTokens
+    // content view on top of `Drawer`
+    public var content: Content
+
+    // configure the behavior of drawer
+    @ObservedObject public var state = DrawerState()
+
+    // configure the apperance of drawer
+    @ObservedObject public var tokens = DrawerTokens()
+
+    @ObservedObject public var viewModel = DrawerAnimationModel()
 
     private let backgroundLayerColor: Color = .black
     private var backgroundLayerOpacity: Double {
@@ -83,11 +100,8 @@ public struct Drawer<Content: View>: View {
     private let percentWidthOfContent: CGFloat = 0.9
     private let percentSnapWidthOfScreen: CGFloat = 0.225
 
-    // content view on top of
-    public var content: Content
-
     // `ivar` to keep track of dragged offset
-    @State private var draggedOffsetWidth: CGFloat?
+    @State internal var draggedOffsetWidth: CGFloat?
     private func computedOffset(screenSize: CGSize) -> CGFloat {
 
         if let draggedOffsetWidth = draggedOffsetWidth {
@@ -96,17 +110,11 @@ public struct Drawer<Content: View>: View {
 
         var width = CGFloat.zero
         let contentWidth = screenSize.width * percentWidthOfContent
-        if !state.isExpanded { // if closed then hidden behind the screen
+        if !viewModel.expand { // if closed then hidden behind the screen
             width = state.presentationDirection == .left ? -contentWidth : contentWidth
         }
 
         return width
-    }
-
-    public init(content: Content) {
-        self.content = content
-        tokens = DrawerTokens()
-        state = DrawerState()
     }
 
     public var body: some View {
@@ -117,6 +125,7 @@ public struct Drawer<Content: View>: View {
                         state.isExpanded.toggle()
                     }
                 }
+
                 content
                     .frame(width: proxy.portraitOrientationAgnosticSize().width * percentWidthOfContent)
                     .shadow(color: tokens.shadowColor.opacity(state.isExpanded ? tokens.shadowOpacity : 0),
@@ -124,7 +133,6 @@ public struct Drawer<Content: View>: View {
                             x: tokens.shadowDepth[0],
                             y: tokens.shadowDepth[1])
                     .offset(x: computedOffset(screenSize: proxy.portraitOrientationAgnosticSize()))
-                    .animation(backgroundLayerAnimation)
 
                 if state.presentationDirection == .left {
                     InteractiveSpacer().onTapGesture {
@@ -136,27 +144,9 @@ public struct Drawer<Content: View>: View {
             .gesture(dragGesture(snapWidth: proxy.portraitOrientationAgnosticSize().width * percentSnapWidthOfScreen))
         }
     }
-
-    private func dragGesture(snapWidth: CGFloat) -> some Gesture {
-        DragGesture()
-            .onChanged { value in
-                let withinDragBounds = state.presentationDirection == .left ? value.translation.width < 0 : value.translation.width > 0
-                if withinDragBounds {
-                    draggedOffsetWidth = value.translation.width
-                }
-            }
-            .onEnded { _ in
-                if let draggedOffsetWidth = draggedOffsetWidth {
-                    if abs(draggedOffsetWidth) < snapWidth {
-                        state.isExpanded = true
-                    } else {
-                        state.isExpanded = false
-                    }
-                    self.draggedOffsetWidth = nil
-                }
-            }
-    }
 }
+
+// MARK: Utilieties + Helpers
 
 extension GeometryProxy {
     func portraitOrientationAgnosticSize() -> CGSize {
@@ -169,163 +159,12 @@ extension GeometryProxy {
     }
 }
 
+// MARK: Composite View
+
 public struct InteractiveSpacer: View {
     public var body: some View {
         ZStack {
             Color.black.opacity(0.001)
         }
-    }
-}
-
-// Mark - Drawer Preview SwiftUI
-
-struct DrawerContent: View {
-    var body: some View {
-        ZStack {
-            Color.red
-            Text("Tap outside to collapse.")
-        }
-    }
-}
-
-struct DrawerPreview: View {
-    var drawer = Drawer(content: DrawerContent())
-    var body: some View {
-        ZStack {
-            NavigationView {
-                EmptyView()
-                    .navigationBarTitle(Text("Drawer Background"))
-                    .navigationBarItems(leading: Button(action: {
-                        drawer.state.isExpanded.toggle()
-                    }, label: {
-                        Image(systemName: "sidebar.left")
-                    })).background(Color.blue)
-            }
-            drawer
-        }
-    }
-}
-
-#if DEBUG
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        DrawerPreview()
-    }
-}
-#endif
-
-// Mark - Drawer UIKit 
-
-/**
- `DrawerHost` is UIKit wrapper required to host UIViewController as content.
- */
-public struct DrawerModal {
-
-    public struct DrawerShimView: View {
-
-        private var drawer: Drawer<DrawerContentViewController>
-
-        public var drawerState: DrawerState {
-            return drawer.state
-        }
-
-        init(contentViewController: UIViewController) {
-            drawer = Drawer(content: DrawerContentViewController(contentViewController: contentViewController))
-        }
-
-        public var body: some View {
-            ZStack {
-                drawer.edgesIgnoringSafeArea(.all)
-            }
-        }
-    }
-
-    public struct DrawerContentViewController: UIViewControllerRepresentable {
-
-        private var contentView: UIViewController
-
-        init(contentViewController: UIViewController) {
-            self.contentView = contentViewController
-        }
-
-        public func makeUIViewController(context: Context) -> UIViewController {
-            return contentView
-        }
-
-        public func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
-
-        public typealias UIViewControllerType = UIViewController
-    }
-}
-
-@objc(MSFDrawerVnext)
-
-/// UIKit wrapper that exposes the SwiftUI Drawer implementation
-open class DrawerVnext: UIHostingController<DrawerModal.DrawerShimView> {
-
-    @objc open var drawerState: DrawerState {
-        return self.rootView.drawerState
-    }
-
-    @objc public init(contentViewController: UIViewController) {
-        super.init(rootView: DrawerModal.DrawerShimView(contentViewController: contentViewController))
-        transitioningDelegate = self
-        drawerState.delegate = self
-        view.backgroundColor = .clear
-        modalPresentationStyle = .overFullScreen
-    }
-
-    @objc required dynamic public init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-    }
-}
-
-extension DrawerVnext: DrawerStateDelegate {
-    // two way binding if state is changed
-    func drawerDidCollapsed() {
-        dismiss(animated: true)
-    }
-}
-
-extension DrawerVnext: UIViewControllerTransitioningDelegate, UIViewControllerAnimatedTransitioning {
-    public func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
-        return 0
-    }
-
-    public func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
-
-        let fromView = transitionContext.viewController(forKey: .from)!.view!
-        let toView = transitionContext.viewController(forKey: .to)!.view!
-
-        let isPresentingDrawer = toView == view
-
-        let drawerView = isPresentingDrawer ? toView : fromView
-
-        // delegate animation to swiftui by changing state
-        if isPresentingDrawer {
-            transitionContext.containerView.addSubview(drawerView)
-            drawerView.frame = UIScreen.main.bounds
-            // Added a static interval as currently swiftui doesn't have an animation completion callback
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.25) { [weak self] in
-                if let strongSelf = self {
-                    strongSelf.drawerState.isExpanded = isPresentingDrawer
-                }
-                transitionContext.completeTransition(true)
-            }
-        } else {
-            self.drawerState.isExpanded = isPresentingDrawer
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.75) {
-                drawerView.removeFromSuperview()
-                transitionContext.completeTransition(true)
-            }
-        }
-    }
-
-    public func animationController(forDismissed dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
-        return self
-    }
-
-    public func animationController(forPresented presented: UIViewController, presenting: UIViewController, source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
-        return self
     }
 }
