+++ ios/FluentUI/DrawerVnext/DrawerPreview.swift
@@ -1,0 +1,44 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import SwiftUI
+
+struct DrawerContent: View {
+    var body: some View {
+        ZStack {
+            Color.red
+            Text("Tap outside to collapse.")
+        }
+    }
+}
+
+struct DrawerPreview: View {
+    var drawer = Drawer(content: DrawerContent())
+    var body: some View {
+        ZStack {
+            NavigationView {
+                EmptyView()
+                    .navigationBarTitle(Text("Drawer Background"))
+                    .navigationBarItems(leading: Button(action: {
+                        drawer.state.isExpanded.toggle()
+                    }, label: {
+                        Image(systemName: "sidebar.left")
+                    })).background(Color.blue)
+            }
+            drawer
+                .didChangedState { value in
+                    print("drawer isExpanded: \(value)")
+                }
+        }
+    }
+}
+
+#if DEBUG
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        DrawerPreview()
+    }
+}
+#endif
